# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NomadBank API
  version: 2.0.0
  description: |
    NomadBank 单用户自托管 API。除初始化、登录和健康检查外，接口均使用
    `nomadbank_session` HttpOnly Cookie 认证。
servers:
  - url: /api/v1
    description: 与 Web 界面同源
tags:
  - name: System
  - name: Session
  - name: Accounts
  - name: Strategies
  - name: Tasks
  - name: Dashboard

paths:
  /health:
    get:
      tags: [System]
      summary: 存活检查
      security: []
      servers:
        - url: /
      responses:
        '200':
          description: 进程正常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /health/ready:
    get:
      tags: [System]
      summary: 就绪检查
      security: []
      servers:
        - url: /
      responses:
        '200':
          description: 数据库可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
        '503':
          $ref: '#/components/responses/Error'
  /setup:
    get:
      tags: [System]
      summary: 查询是否已经初始化
      security: []
      responses:
        '200':
          description: 初始化状态
          content:
            application/json:
              schema:
                type: object
                required: [initialized]
                properties:
                  initialized:
                    type: boolean
    post:
      tags: [System]
      summary: 创建唯一所有者
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupInput'
      responses:
        '201':
          description: 初始化完成并设置会话 Cookie
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
  /session:
    post:
      tags: [Session]
      summary: 登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: 登录成功
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Session]
      summary: 退出当前会话
      responses:
        '204':
          description: 已退出
          headers:
            Set-Cookie:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Error'
  /me:
    get:
      tags: [Session]
      summary: 获取所有者资料
      responses:
        '200':
          description: 当前所有者
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '401':
          $ref: '#/components/responses/Error'
    put:
      tags: [Session]
      summary: 更新显示名称和时区
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [display_name, timezone]
              properties:
                display_name:
                  type: string
                  maxLength: 80
                  description: 可为空字符串，用于清除显示名称
                timezone:
                  type: string
                  minLength: 1
                  examples: [Asia/Shanghai]
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
  /me/password:
    put:
      tags: [Session]
      summary: 修改密码并撤销其他会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                  format: password
                new_password:
                  type: string
                  format: password
                  minLength: 10
                  description: 至少 10 个 Unicode 字符，UTF-8 编码后最多 72 字节
      responses:
        '204':
          description: 密码已更新
          headers:
            Set-Cookie:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
  /accounts:
    get:
      tags: [Accounts]
      summary: 获取全部银行账户
      responses:
        '200':
          description: 账户列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Error'
    post:
      tags: [Accounts]
      summary: 创建银行账户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountInput'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
  /accounts/{id}:
    parameters:
      - $ref: '#/components/parameters/ID'
    get:
      tags: [Accounts]
      summary: 获取账户
      responses:
        '200':
          description: 账户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    put:
      tags: [Accounts]
      summary: 更新账户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountInput'
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Accounts]
      summary: 删除未被任务引用的账户
      responses:
        '204':
          description: 已删除
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /strategies:
    get:
      tags: [Strategies]
      summary: 获取全部策略
      responses:
        '200':
          description: 策略列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Strategy'
        '401':
          $ref: '#/components/responses/Error'
    post:
      tags: [Strategies]
      summary: 创建策略
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyInput'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
  /strategies/{id}:
    parameters:
      - $ref: '#/components/parameters/ID'
    get:
      tags: [Strategies]
      summary: 获取策略
      responses:
        '200':
          description: 策略
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    put:
      tags: [Strategies]
      summary: 更新策略
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyInput'
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Strategies]
      summary: 删除策略
      responses:
        '204':
          description: 已删除，历史批次保留策略名称
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /task-batches:
    get:
      tags: [Tasks]
      summary: 获取任务批次
      responses:
        '200':
          description: 批次列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskBatch'
        '401':
          $ref: '#/components/responses/Error'
    post:
      tags: [Tasks]
      summary: 生成任务批次
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [strategy_id]
              properties:
                strategy_id:
                  type: integer
                  format: int64
                  minimum: 1
                  maximum: 9223372036854775807
                group_name:
                  type: string
                  default: ''
                cycles:
                  type: integer
                  minimum: 1
                  maximum: 24
                  default: 4
      responses:
        '201':
          description: 已生成
          content:
            application/json:
              schema:
                type: object
                required: [batch, tasks]
                properties:
                  batch:
                    $ref: '#/components/schemas/TaskBatch'
                  tasks:
                    type: integer
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
  /task-batches/{id}:
    parameters:
      - $ref: '#/components/parameters/ID'
    delete:
      tags: [Tasks]
      summary: 删除任务批次及其任务
      responses:
        '204':
          description: 已删除
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /tasks:
    get:
      tags: [Tasks]
      summary: 分页获取任务
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed]
        - name: batch_id
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 9223372036854775807
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100000
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 任务分页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPage'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /tasks/{id}/complete:
    parameters:
      - $ref: '#/components/parameters/ID'
    post:
      tags: [Tasks]
      summary: 幂等地完成任务
      responses:
        '200':
          description: 已完成；重复调用返回原完成时间
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /dashboard:
    get:
      tags: [Dashboard]
      summary: 获取仪表盘摘要
      responses:
        '200':
          description: 摘要、即将执行与最近完成任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/Error'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nomadbank_session
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
        maximum: 9223372036854775807
  responses:
    Error:
      description: 请求失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    SetupInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 40
        password:
          type: string
          format: password
          minLength: 10
          description: 至少 10 个 Unicode 字符，UTF-8 编码后最多 72 字节
        display_name:
          type: string
          maxLength: 80
        timezone:
          type: string
          default: Asia/Shanghai
          description: 省略或传空字符串时使用 Asia/Shanghai
          examples: [Asia/Shanghai]
    Owner:
      type: object
      required: [username, display_name, timezone]
      properties:
        username:
          type: string
        display_name:
          type: string
          maxLength: 80
        timezone:
          type: string
    AccountInput:
      type: object
      required: [name, group_name, active]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        group_name:
          type: string
          maxLength: 50
        active:
          type: boolean
    Account:
      allOf:
        - $ref: '#/components/schemas/AccountInput'
        - type: object
          required: [id, created_at, updated_at]
          properties:
            id:
              type: integer
              format: int64
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    StrategyInput:
      type: object
      required:
        - name
        - interval_min_days
        - interval_max_days
        - time_start_minutes
        - time_end_minutes
        - skip_weekends
        - amount_min_cents
        - amount_max_cents
        - daily_limit
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        interval_min_days:
          type: integer
          minimum: 1
          maximum: 365
        interval_max_days:
          type: integer
          minimum: 1
          maximum: 365
        time_start_minutes:
          type: integer
          minimum: 0
          maximum: 1439
        time_end_minutes:
          type: integer
          minimum: 1
          maximum: 1440
        skip_weekends:
          type: boolean
        amount_min_cents:
          type: integer
          format: int64
          minimum: 1
          maximum: 100000000
        amount_max_cents:
          type: integer
          format: int64
          minimum: 1
          maximum: 100000000
        daily_limit:
          type: integer
          minimum: 1
          maximum: 100
    Strategy:
      allOf:
        - $ref: '#/components/schemas/StrategyInput'
        - type: object
          required: [id, created_at, updated_at]
          properties:
            id:
              type: integer
              format: int64
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    TaskBatch:
      type: object
      required: [id, strategy_id, strategy_name, group_name, cycle_count, task_count, created_at]
      properties:
        id:
          type: integer
          format: int64
        strategy_id:
          type: [integer, 'null']
          format: int64
        strategy_name:
          type: string
        group_name:
          type: string
        cycle_count:
          type: integer
        task_count:
          type: integer
        created_at:
          type: string
          format: date-time
    Task:
      type: object
      required:
        - id
        - batch_id
        - cycle_no
        - scheduled_at
        - from_account_id
        - from_account_name
        - to_account_id
        - to_account_name
        - amount_cents
        - status
        - completed_at
        - created_at
      properties:
        id:
          type: integer
          format: int64
        batch_id:
          type: integer
          format: int64
        cycle_no:
          type: integer
        scheduled_at:
          type: string
          format: date-time
        from_account_id:
          type: integer
          format: int64
        from_account_name:
          type: string
        to_account_id:
          type: integer
          format: int64
        to_account_name:
          type: string
        amount_cents:
          type: integer
          format: int64
        status:
          type: string
          enum: [pending, completed]
        completed_at:
          type: [string, 'null']
          format: date-time
        created_at:
          type: string
          format: date-time
    TaskPage:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        total:
          type: integer
          format: int64
        page:
          type: integer
        page_size:
          type: integer
    Dashboard:
      type: object
      required:
        - total_accounts
        - active_accounts
        - pending_tasks
        - completed_tasks
        - strategies
        - upcoming
        - recent
      properties:
        total_accounts:
          type: integer
          format: int64
        active_accounts:
          type: integer
          format: int64
        pending_tasks:
          type: integer
          format: int64
        completed_tasks:
          type: integer
          format: int64
        strategies:
          type: integer
          format: int64
        upcoming:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        recent:
          type: array
          items:
            $ref: '#/components/schemas/Task'

security:
  - cookieAuth: []